+++ app/src/components/storefront/StorefrontNavbar.tsx
@@ -1,65 +1,9 @@
-import { Link, useNavigate, useSearch } from "@tanstack/react-router";
+import { Link } from "@tanstack/react-router";
 import KFKLogo from "@/assets/kfk-logo.png";
-import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
-import {
-  DropdownMenu,
-  DropdownMenuItem,
-  DropdownMenuContent,
-  DropdownMenuLabel,
-  DropdownMenuRadioGroup,
-  DropdownMenuRadioItem,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import {
-  AdjustmentsHorizontalIcon,
-  ArrowTopRightOnSquareIcon,
-  MagnifyingGlassIcon,
-  ShoppingCartIcon,
-} from "../icons";
-
-export type SortOption =
-  | "age-asc"
-  | "age-desc"
-  | "gifts-asc"
-  | "gifts-desc"
-  | undefined;
+import { ArrowTopRightOnSquareIcon, ShoppingCartIcon } from "../icons";
 
 export function StorefrontNavbar() {
-  const navigate = useNavigate();
-  // Reads search/sort param — returns undefined on routes that don't define it
-  const search = useSearch({ strict: false }) as {
-    search?: string;
-    sort?: SortOption;
-  };
-  const searchValue = search?.search ?? "";
-  const sortValue = search?.sort ?? "";
-
-  const handleSearch = (value: string) => {
-    navigate({
-      // @ts-ignore — search param is defined per-route
-      search: (prev: any) => ({ ...prev, search: value || undefined }),
-    });
-  };
-
-  const handleSort = (value: string) => {
-    navigate({
-      // @ts-ignore — router search params typing mismatch
-      search: (prev: any) => ({
-        ...prev,
-        sort: value || undefined,
-      }),
-    });
-  };
-
-  const sortLabel: Record<string, string> = {
-    "age-asc": "Age: Youngest → Oldest",
-    "age-desc": "Age: Oldest → Youngest",
-    "gifts-asc": "Gifts Fulfilled: Least → Most",
-    "gifts-desc": "Gifts Fulfilled: Most → Least",
-  };
-
   return (
     <div className="mx-8">
       <div className="px-4 flex items-center justify-between">
@@ -115,63 +59,6 @@ export function StorefrontNavbar() {
       </div>
 
       <div className="-mx-8 my-4 bg-sidebar-ring py-[0.5px]"></div>
-
-      <div className="px-4 pb-3 flex items-center gap-3">
-        <DropdownMenu>
-          <DropdownMenuTrigger asChild>
-            <Button
-              variant="outline"
-              className="border-sidebar-ring text-muted-foreground hover:bg-transparent whitespace-nowrap"
-            >
-              <AdjustmentsHorizontalIcon />
-              {sortValue ? sortLabel[sortValue] : "Filters"}
-            </Button>
-          </DropdownMenuTrigger>
-          <DropdownMenuContent className="w-56">
-            <DropdownMenuLabel>Sort By</DropdownMenuLabel>
-            <DropdownMenuSeparator />
-            <DropdownMenuRadioGroup
-              value={sortValue}
-              onValueChange={handleSort}
-            >
-              <DropdownMenuRadioItem value="age-asc">
-                Age: Youngest → Oldest
-              </DropdownMenuRadioItem>
-              <DropdownMenuRadioItem value="age-desc">
-                Age: Oldest → Youngest
-              </DropdownMenuRadioItem>
-              <DropdownMenuSeparator />
-              <DropdownMenuRadioItem value="gifts-asc">
-                Gifts Fulfilled: Least → Most
-              </DropdownMenuRadioItem>
-              <DropdownMenuRadioItem value="gifts-desc">
-                Gifts Fulfilled: Most → Least
-              </DropdownMenuRadioItem>
-            </DropdownMenuRadioGroup>
-            {sortValue && (
-              <>
-                <DropdownMenuSeparator />
-                <DropdownMenuItem
-                  onClick={() => handleSort("")}
-                  className="text-muted-foreground"
-                >
-                  Clear filter
-                </DropdownMenuItem>
-              </>
-            )}
-          </DropdownMenuContent>
-        </DropdownMenu>
-
-        <div className="relative w-full">
-          <MagnifyingGlassIcon className="absolute left-3 top-1/2 -translate-y-1/2 h-4 w-4 text-muted-foreground" />
-          <Input
-            placeholder="Search"
-            className="pl-9 border-sidebar-ring"
-            value={searchValue}
-            onChange={(e) => handleSearch(e.target.value)}
-          />
-        </div>
-      </div>
     </div>
   );
 }
